requirements.txt
flask
transformers<5.0.0
torch
newspaper3k
lxml
requests
lxml_html_clean
